% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_list}
\alias{clean_list}
\title{Clean list of Swiss Air Pollutant dataframe}
\usage{
clean_list(loaded_file, pollutant)
}
\arguments{
\item{loaded_file}{A \code{list} list containing all the dataframes (return by load_data())}

\item{pollutant}{A \code{char} vector containing names of pollutant}
}
\value{
A \code{list} a cleaned and labelled list
}
\description{
Takes an untidied list of dataframe and cleans it. Cleans the locations' names and variables' type.
}
\author{
Anna Alfieri, Ana Lucy Bejarano Montalvo, Saphir Kwan, Erika Lardo, Clément Perez
}
