% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_names}
\alias{clean_names}
\title{Clean Swiss Location Names}
\usage{
clean_names(name_vec)
}
\arguments{
\item{name_vec}{A \code{char} vector containing the names to clean}
}
\value{
A \code{vector} containing the cleaned names
}
\description{
Cleans names of Swiss location by removing all accent and replacing dashes by underscore.
}
\author{
Anna Alfieri, Ana Lucy Bejarano Montalvo, Saphir Kwan, Erika Lardo, Clément Perez
}
