% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_location_name}
\alias{get_location_name}
\title{Retrieve Swiss Location Names}
\usage{
get_location_name(loaded_file)
}
\arguments{
\item{name_vec}{A \code{list} list containing the datasets as returned by clean_file()}
}
\value{
A \code{vector} contains the locations' names
}
\description{
Retrieve the names of the Swiss locations present in the data
}
\author{
Anna Alfieri, Ana Lucy Bejarano Montalvo, Saphir Kwan, Erika Lardo, Clément Perez
}
