% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_data}
\alias{load_data}
\title{Load Swiss Air Pollutant Data}
\usage{
load_data(path, pollutant, file_ext)
}
\arguments{
\item{path}{A \code{char} path to locate the target files}

\item{pollutant}{A \code{char} vector containing the names of the pollutants you want to load}

\item{file_ext}{A\code{char} string containing the second part of the target files' names}
}
\value{
A \code{list} containing the loaded data
}
\description{
Load into a list several dataset as retrieved from \url{https://www.bafu.admin.ch/bafu/en/home/topics/air/state/data/data-query-nabel.html}.
The function \code{load_data()} is similar to \code{load_data2()}, however it is adapted for data separated by comas (,).
Please name your datasets following this pattern: \code{pollutant_date.csv}  where "pollutant" is the name of the pollutant
and "date" the concerned months.
}
\author{
Anna Alfieri, Ana Lucy Bejarano Montalvo, Saphir Kwan, Erika Lardo, Clement Perez
}
