% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{run_demo}
\alias{run_demo}
\title{Run Demo}
\usage{
run_demo()
}
\description{
Allows to run a demo of our shiny app which presents visualizations of air pollution in Switzerland
}
\author{
Anna Alfieri, Ana Lucy Bejarano Montalvo, Saphir Kwan, Erika Lardo, Clement Perez
}
